+++ lib/screens/screens.dart
@@ -1,34 +1,34 @@
-export 'package:vende_facil/screens/abono_detalle.dart';
-export 'package:vende_facil/screens/tarjetas_screen.dart';
-export 'package:vende_facil/screens/suscripcion_screen.dart';
+export 'package:vende_facil/screens/apartados/abono_detalle.dart';
+export 'package:vende_facil/screens/suscripcion/tarjetas_screen.dart';
+export 'package:vende_facil/screens/suscripcion/suscripcion_screen.dart';
 export 'package:vende_facil/screens/menu.dart';
-export 'package:vende_facil/screens/agregar_abono.dart';
-export 'package:vende_facil/screens/agregar_categoria.dart';
-export 'package:vende_facil/screens/agregar_cliente.dart';
-export 'package:vende_facil/screens/agregar_descuento.dart';
-export 'package:vende_facil/screens/agregar_empresa.dart';
-export 'package:vende_facil/screens/agregar_producto.dart';
-export 'package:vende_facil/screens/configuracion_screen.dart';
+export 'package:vende_facil/screens/apartados/agregar_abono.dart';
+export 'package:vende_facil/screens/categorias/agregar_categoria.dart';
+export 'package:vende_facil/screens/clientes/agregar_cliente.dart';
+export 'package:vende_facil/screens/descuentos/agregar_descuento.dart';
+export 'package:vende_facil/screens/empresa/agregar_empresa.dart';
+export 'package:vende_facil/screens/productos/agregar_producto.dart';
+export 'package:vende_facil/screens/configuracion/configuracion_screen.dart';
 export 'package:vende_facil/screens/error_screen.dart';
-export 'package:vende_facil/screens/historial_screen.dart';
+export 'package:vende_facil/screens/ventas/historial_screen.dart';
 export 'package:vende_facil/screens/home_screen.dart';
-export 'package:vende_facil/screens/lista_categorias.dart';
-export 'package:vende_facil/screens/lista_clientes.dart';
-export 'package:vende_facil/screens/lista_descuentos.dart';
-export 'package:vende_facil/screens/lista_productos.dart';
-export 'package:vende_facil/screens/login.dart';
-export 'package:vende_facil/screens/registro_screen.dart';
-export 'package:vende_facil/screens/venta_detalle_screen.dart';
+export 'package:vende_facil/screens/categorias/lista_categorias.dart';
+export 'package:vende_facil/screens/clientes/lista_clientes.dart';
+export 'package:vende_facil/screens/descuentos/lista_descuentos.dart';
+export 'package:vende_facil/screens/productos/lista_productos.dart';
+export 'package:vende_facil/screens/usuarios/login.dart';
+export 'package:vende_facil/screens/usuarios/registro_screen.dart';
+export 'package:vende_facil/screens/ventas/venta_detalle_screen.dart';
 export 'package:vende_facil/screens/search_screen.dart';
-export 'package:vende_facil/screens/agregar_tarjeta.dart';
-export 'package:vende_facil/screens/detalles_venta.dart';
-export 'package:vende_facil/screens/perfil_screen.dart';
-export 'package:vende_facil/screens/config_apartado_screen.dart';
-export 'package:vende_facil/screens/search_screen_abonos.dart';
-export 'package:vende_facil/screens/agregar_empleado_screen.dart';
-export 'package:vende_facil/screens/lista_empleados.dart';
-export 'package:vende_facil/screens/perfil_empleado_screen.dart';
-export 'package:vende_facil/screens/venta_screen.dart';
-export 'package:vende_facil/screens/apartado_screen.dart';
+export 'package:vende_facil/screens/suscripcion/agregar_tarjeta.dart';
+export 'package:vende_facil/screens/ventas/detalles_venta.dart';
+export 'package:vende_facil/screens/usuarios/perfil_screen.dart';
+export 'package:vende_facil/screens/configuracion/config_apartado_screen.dart';
+export 'package:vende_facil/screens/apartados/search_screen_abonos.dart';
+export 'package:vende_facil/screens/empresa/agregar_empleado_screen.dart';
+export 'package:vende_facil/screens/empresa/lista_empleados.dart';
+export 'package:vende_facil/screens/empresa/perfil_empleado_screen.dart';
+export 'package:vende_facil/screens/ventas/venta_screen.dart';
+export 'package:vende_facil/screens/apartados/apartado_screen.dart';
 export 'package:vende_facil/screens/splash_screen.dart';
-export 'package:vende_facil/screens/planes_screen.dart';
+export 'package:vende_facil/screens/suscripcion/planes_screen.dart';
